Completed exercise 2
main clarification was needed in defining when to use classes vs when to use attributes. Learned an ID is specific to a single element while one can assign classes to multiple elements on a page or throughout site.

IDs are unique to the page, so better for identification. Classes are supposed to be used to group similar style things together.

Apply something to ONE SPECIFIC div (element) use an id because it will take priority (more specific)