Complete exercise 6
Many ways to complete. Solved mine using what I believe is an easy to read specificity from top to bottom. I got the right output and it is easy for me to read, but I have no idea if the syntax is correct.

This exercise was challenging for me because I still do not understand the specifics of the hierarchical order of CSS.

Continue to practice with CSS. More mileage is obviously the best feedback for not understanding.

All things considered, and upon closer inspection, my solution is quite similar to the solution.css.