Complete exercise 05
used flex column to try to align message correctly.

My method breaks the message bocks into 3 columns: icon, text/button/header, and x button.

I createed a class to group the middle column and another class to group the continue/cancel buttons.

The only downside I saw with this is that the text will not bleed into the far right margin (below the x button). It does still look clean, though and requires roughly the same amount of code as the solution.css.

The TOP solution groups the header, close button, and the two bottom buttons together in the 'container' class.

It does NOT include the icon class in the container. This allows the icon to remain aligned to .modal.

It aligns content in the header with space-between.